Moving token creation to a utility function
